Update postgresql cookbook
There is a problem with installing PostgreSQL developer tools
version 9.1 which is compatible with PostgreSQL server 9.1. The
cookbook checks if server and developer tools versions are the
same. If so, then the PostgreSQL version number is defaulted based
on the Centos OS version which is PostgreSQL 8.4 for CentosOS 6.7.

This version conditional check has been removed to allow for 9.1
version of the server and developer tools to be installed in
Centos 6.7.